server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
mysql-connector-python==8.3.0
jinja2==3.1.3
python-multipart==0.0.9